Refactor juniper_integration_tests and test juniper_codegen_tests only on nightly
