+++ tests/codegen/Cargo.toml
@@ -5,10 +5,11 @@ edition = "2021"
 publish = false
 
 [dependencies]
-futures = "0.3.1"
-juniper = { path = "../../juniper" }
+rustversion = "1.0"
 
 [dev-dependencies]
 serde_json = "1.0"
 tokio = { version = "1.0", features = ["rt", "time", "macros"] }
-trybuild = "1.0.25"
+trybuild = "1.0.63"
+futures = "0.3.1"
+juniper = { path = "../../juniper" }
