+++ tests/integration/Cargo.toml
@@ -4,14 +4,12 @@ version = "0.0.0"
 edition = "2021"
 publish = false
 
-[dependencies]
+[dev-dependencies]
 chrono = "0.4"
 derive_more = "0.99"
 futures = "0.3"
 juniper = { path = "../../juniper" }
 juniper_subscriptions = { path = "../../juniper_subscriptions" }
-
-[dev-dependencies]
 async-trait = "0.1.39"
 fnv = "1.0"
 serde = { version = "1.0", features = ["derive"] }
