--- tests/integration/src/codegen/mod.rs
@@ -1,15 +1,0 @@
-mod derive_object_with_raw_idents;
-mod enum_derive;
-mod input_object_derive;
-mod interface_attr_struct;
-mod interface_attr_trait;
-mod interface_derive;
-mod object_attr;
-mod object_derive;
-mod scalar_attr_derive_input;
-mod scalar_attr_type_alias;
-mod scalar_derive;
-mod scalar_value_derive;
-mod subscription_attr;
-mod union_attr;
-mod union_derive;
