+++ tests/integration/tests/custom_scalar.rs
@@ -1,95 +1,15 @@
-use std::{fmt, pin::Pin};
+pub mod common;
 
+use std::pin::Pin;
+
+use common::MyScalarValue;
 use futures::{stream, Stream};
 use juniper::{
     execute, graphql_input_value, graphql_object, graphql_scalar, graphql_subscription,
     graphql_vars,
     parser::{ParseError, ScalarToken, Token},
-    serde::{de, Deserialize, Deserializer, Serialize},
-    EmptyMutation, FieldResult, InputValue, Object, ParseScalarResult, RootNode, ScalarValue,
-    Value, Variables,
+    EmptyMutation, FieldResult, InputValue, Object, ParseScalarResult, RootNode, Value, Variables,
 };
-
-#[derive(Clone, Debug, PartialEq, ScalarValue, Serialize)]
-#[serde(untagged)]
-pub(crate) enum MyScalarValue {
-    #[value(as_float, as_int)]
-    Int(i32),
-    Long(i64),
-    #[value(as_float)]
-    Float(f64),
-    #[value(as_str, as_string, into_string)]
-    String(String),
-    #[value(as_bool)]
-    Boolean(bool),
-}
-
-impl<'de> Deserialize<'de> for MyScalarValue {
-    fn deserialize<D: Deserializer<'de>>(de: D) -> Result<Self, D::Error> {
-        struct Visitor;
-
-        impl<'de> de::Visitor<'de> for Visitor {
-            type Value = MyScalarValue;
-
-            fn expecting(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-                f.write_str("a valid input value")
-            }
-
-            fn visit_bool<E: de::Error>(self, b: bool) -> Result<Self::Value, E> {
-                Ok(MyScalarValue::Boolean(b))
-            }
-
-            fn visit_i32<E: de::Error>(self, n: i32) -> Result<Self::Value, E> {
-                Ok(MyScalarValue::Int(n))
-            }
-
-            fn visit_i64<E: de::Error>(self, b: i64) -> Result<Self::Value, E> {
-                if b <= i64::from(i32::MAX) {
-                    self.visit_i32(b.try_into().unwrap())
-                } else {
-                    Ok(MyScalarValue::Long(b))
-                }
-            }
-
-            fn visit_u32<E: de::Error>(self, n: u32) -> Result<Self::Value, E> {
-                if n <= i32::MAX as u32 {
-                    self.visit_i32(n.try_into().unwrap())
-                } else {
-                    self.visit_u64(n.into())
-                }
-            }
-
-            fn visit_u64<E: de::Error>(self, n: u64) -> Result<Self::Value, E> {
-                if n <= i64::MAX as u64 {
-                    self.visit_i64(n.try_into().unwrap())
-                } else {
-                    // Browser's `JSON.stringify()` serializes all numbers
-                    // having no fractional part as integers (no decimal point),
-                    // so we must parse large integers as floating point,
-                    // otherwise we would error on transferring large floating
-                    // point numbers.
-                    // TODO: Use `FloatToInt` conversion once stabilized:
-                    //       https://github.com/rust-lang/rust/issues/67057
-                    Ok(MyScalarValue::Float(n as f64))
-                }
-            }
-
-            fn visit_f64<E: de::Error>(self, f: f64) -> Result<Self::Value, E> {
-                Ok(MyScalarValue::Float(f))
-            }
-
-            fn visit_str<E: de::Error>(self, s: &str) -> Result<Self::Value, E> {
-                self.visit_string(s.into())
-            }
-
-            fn visit_string<E: de::Error>(self, s: String) -> Result<Self::Value, E> {
-                Ok(MyScalarValue::String(s))
-            }
-        }
-
-        de.deserialize_any(Visitor)
-    }
-}
 
 #[graphql_scalar(with = long, scalar = MyScalarValue)]
 type Long = i64;
