+++ tests/integration/tests/derive_object_with_raw_idents.rs
