+++ tests/integration/tests/input_object_derive.rs
@@ -1,11 +1,12 @@
 //! Tests for `#[derive(GraphQLInputObject)]` macro.
+pub mod common;
 
 use juniper::{
     execute, graphql_object, graphql_value, graphql_vars, parser::SourcePosition, GraphQLError,
     GraphQLInputObject, RuleError,
 };
 
-use crate::util::schema;
+use common::util::schema;
 
 mod trivial {
     use super::*;
