+++ tests/integration/tests/inside_macro.rs
