+++ tests/integration/tests/issue_371.rs
