+++ tests/integration/tests/issue_372.rs
