+++ tests/integration/tests/issue_798.rs
