+++ tests/integration/tests/issue_922.rs
