+++ tests/integration/tests/issue_925.rs
