+++ tests/integration/tests/issue_945.rs
