+++ tests/integration/tests/object_derive.rs
@@ -1,4 +1,5 @@
 //! Tests for `#[derive(GraphQLObject)]` macro.
+pub mod common;
 
 use juniper::{
     execute, graphql_object, graphql_value, graphql_vars, DefaultScalarValue, EmptyMutation,
@@ -851,7 +852,7 @@ mod explicit_scalar {
 }
 
 mod custom_scalar {
-    use crate::custom_scalar::MyScalarValue;
+    use common::MyScalarValue;
 
     use super::*;
