+++ tests/integration/tests/scalar_value_derive.rs
